Finally touched up how the operations will look like

+++ include/DataStructures/Dijkstra.hpp
@@ -21,7 +21,7 @@ public:
 	bool del_edge(int u, int v, int w);
 	std::vector<int> get_vertices();
 	std::vector<DijkstraEdge> get_edges();
-	std::vector<std::pair<int,long long>> run_dijkstra(int u, int v); 
+	std::vector<std::pair<int, int>> run_dijkstra(int u, int v);
 	// get all the vertices in order of visited when finding the shortest path from u to v
 	std::vector<int> get_shortest_path(int u, int v);
 	// get the shortest path from u to v
